requirements.txt
opencv-python==4.8.1.78
face-recognition==1.3.0
deepface==0.0.79
numpy==1.24.3
Pillow==10.0.0
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0
schedule==1.2.0
twilio==8.5.0
Flask==3.0.0
dlib==19.24.2
cmake==3.27.5
